% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFunStats.R
\name{c_fun_stats}
\alias{c_fun_stats}
\title{c_fun_stats}
\usage{
c_fun_stats(x, y, full = FALSE, log.p = FALSE)
}
\arguments{
\item{x}{(factor, integer or numeric) An ordinal categorical independent variable. An ordered factor is recommended,
with pre-defined ordering scheme as Y and X are sorted w.r.t X.}

\item{y}{(factor, integer or numeric) A categorical dependent variable.}

\item{full}{(logical) Disables trimming and returns the compressed global pattern. Default: FALSE}

\item{log.p}{(logical) If natural log of the p-value should be returned. Default: FALSE}
}
\value{
Compact functional statistics for the most compact pattern.
A list with class "\code{htest}" containing the following components:
\describe{
 \item{\code{statistic}}{the functional chi-squared statistic}

 \item{\code{parameter}}{degrees of freedom for the functional chi-squared statistic}

 \item{\code{p.value}}{p-value of the functional test computed by an asymptotic chi-squared distribution}

 \item{\code{estimate}}{an estimate of function index between 0 and 1.
The value of 1 indicates a strictly mathematical function. It is
asymmetrical with respect to transpose of the input contingency table,
different from the symmetrical Cramer's V based on the Pearson's
chi-squared test statistic.}
}
}
\description{
Computes the most compact contingency table and calculates the compact functional statistics
}
\examples{
library(DescTools)
xy_table = matrix(c( 20,20,20,
                     20,20,20,
                     20,20,20,
                     10,0,0,
                     0,10,0,
                     0,0,10,
                     0,10,0,
                     10,0,0,
                     20,20,20,
                     20,20,20,
                     20,20,20 ), nrow=11, ncol=3, byrow = TRUE)
 xy = Untable(xy_table)
 x = xy$Var1
 y = xy$Var2
 print(c_fun_stats(x, y))
}
\keyword{Compact}
\keyword{Functional}
\keyword{Local}
\keyword{dependency}
\keyword{pattern,}
\keyword{representation,}
