% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_table.R
\name{construct_compact_table}
\alias{construct_compact_table}
\title{construct_compact_table}
\usage{
construct_compact_table(x, y, full = FALSE)
}
\arguments{
\item{x}{(factor, integer or numeric) An ordinal categorical independent variable. An ordered factor is recommended,
with pre-defined ordering scheme as y and x are sorted w.r.t x.}

\item{y}{(factor, integer or numeric) A categorical dependent variable.}

\item{full}{(logical) Disables trimming and returns the compressed global pattern. Default: FALSE}
}
\value{
\code{ctable}: most compact cross cross-tabular representation of \code{x} and \code{y}.
}
\description{
Employs dynamic programming to trim and compress a pattern to its most compact cross-tabular representation
}
\examples{
library(DescTools)
xy_table = matrix(c( 20,20,20,
                     20,20,20,
                     20,20,20,
                     10,0,0,
                     0,10,0,
                     0,0,10,
                     0,10,0,
                     10,0,0,
                     20,20,20,
                     20,20,20,
                     20,20,20 ), nrow=11, ncol=3, byrow = TRUE)
 xy = Untable(xy_table)
 x = xy$Var1
 y = xy$Var2
 print(construct_compact_table(x, y))
}
\keyword{Compact}
\keyword{Dynamic}
\keyword{Local}
\keyword{Trimming,}
\keyword{pattern.}
\keyword{programming,}
\keyword{representation,}
